# server/routes/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Employee API",
    "version": "1.0.0",
    "description": "Employee management API"
  },
  "paths": {
    "/api/employees/{empId}": {
      "get": {
        "summary": "Get employee by ID",
        "operationId": "findEmployeeById",
        "parameters": [
          {
            "name": "empId",
            "in": "path",
            "required": true,
            "description": "Employee ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": { "description": "Bad Request" },
          "404": { "description": "Not Found" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/api/employees/{empId}/tasks": {
      "get": {
        "summary": "Get all tasks for an employee",
        "operationId": "findAllTasksForEmployee",
        "parameters": [
          {
            "name": "empId",
            "in": "path",
            "required": true,
            "description": "Employee ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": { "description": "Bad Request" },
          "404": { "description": "Not Found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "post": {
        "summary": "Create a new task for an employee",
        "operationId": "createTaskForEmployee",
        "parameters": [
          {
            "name": "empId",
            "in": "path",
            "required": true,
            "description": "Employee ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "task": {
                    "type": "object",
                    "properties": {
                      "text": {
                        "type": "string"
                      },
                      "category": {
                        "type": "object",
                        "properties": {
                          "categoryName": { "type": "string" },
                          "backgroundColor": { "type": "string" }
                        },
                        "required": ["categoryName", "backgroundColor"]
                      }
                    },
                    "required": ["text", "category"]
                  }
                },
                "required": ["task"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": { "description": "Bad Request" },
          "404": { "description": "Not Found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "put": {
        "summary": "Update tasks for an employee",
        "operationId": "updateTasksForEmployee",
        "parameters": [
          {
            "name": "empId",
            "in": "path",
            "required": true,
            "description": "Employee ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "todo": {
                    "type": "array",
                    "items": { "type": "object" }
                  },
                  "done": {
                    "type": "array",
                    "items": { "type": "object" }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": { "description": "Successfully updated tasks" },
          "400": { "description": "Bad Request" },
          "404": { "description": "Not Found" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/api/employees/{empId}/tasks/{taskId}": {
      "delete": {
        "summary": "Delete a task for an employee",
        "operationId": "deleteTaskForEmployee",
        "parameters": [
          {
            "name": "empId",
            "in": "path",
            "required": true,
            "description": "Employee ID",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "description": "Task ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": { "description": "Successfully deleted" },
          "400": { "description": "Bad Request" },
          "404": { "description": "Not Found" },
          "500": { "description": "Internal Server Error" }
        }
      }
    }
  }
}